ubumtu 20.04 lts

+++ QSDFT_FMT_Slit_alpha_v.0.1.0_float/run.sh
@@ -19,8 +19,18 @@ for w in "${items[@]}"; do
 	echo "Slit width = ${H} [nm]"
 	sed -i "s/XXX/${H}/g" parameters.txt
 	./qsdft_fmt_slit.exe
-	mv PP0_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
-	mv PP0_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	if [ -e PP0_vs_Vgamma_data_vs.txt ]; then
+	  mv PP0_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv PP0_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
+	if [ -e Pa_vs_Vgamma_data_vs.txt ]; then
+	  mv Pa_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv Pa_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
+	if [ -e atm_vs_Vgamma_data_vs.txt ]; then
+	  mv atm_vs_Vgamma_data_vs.txt ${w}_data_vs.txt
+	  mv atm_vs_Vgamma_data_ls.txt ${w}_data_ls.txt
+	fi
 	cp ${w}_data_vs.txt ./results/${w}_data_vs.txt
 	cp ${w}_data_ls.txt ./results/${w}_data_ls.txt
 	rm parameters.txt
